fix double pasting issue
